requirements.txt
aniso8601==9.0.1
anyascii==0.3.2
asgiref==3.6.0
attrs==22.2.0
beautifulsoup4==4.11.2
boto3==1.26.133
botocore==1.29.133
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==40.0.2
Django==4.1.7
django-cors-headers==3.14.0
django-environ==0.9.0
django-filter==22.1
django-modelcluster==6.0
django-permissionedforms==0.1
django-rest-framework==0.1.0
django-storages==1.13.2
django-taggit==3.1.0
django-treebeard==4.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
draftjs-exporter==2.1.7
drf-spectacular==0.26.1
drf-spectacular-sidecar==2023.4.1
environ==1.0
et-xmlfile==1.1.0
graphene==3.2.2
graphene-django==3.0.2
graphql-core==3.2.3
graphql-relay==3.2.0
gunicorn==20.1.0
html5lib==1.1
idna==3.4
inflection==0.5.1
jmespath==1.0.1
josepy==1.13.0
jsonschema==4.17.3
l18n==2021.3
mozilla-django-oidc==3.0.0
openpyxl==3.1.2
Pillow==9.5.0
promise==2.3
psycopg2-binary==2.9.6
pycparser==2.21
PyJWT==2.6.0
pyOpenSSL==23.1.1
pyrsistent==0.19.3
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
requests==2.28.2
s3transfer==0.6.1
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.3
telepath==0.3
text-unidecode==1.3
uritemplate==4.1.1
urllib3==1.26.15
wagtail==4.2.2
webencodings==0.5.1
Willow==1.4.1
yandex-geocoder==3.0.1
